+++ example/trajectory_example.cpp
@@ -2,45 +2,78 @@
 #include <Elite/DataType.hpp>
 #include <Elite/EliteDriver.hpp>
 #include <Elite/RtsiIOInterface.hpp>
+#include <Elite/Log.hpp>
 
+#include <boost/program_options.hpp>
 #include <iostream>
 #include <memory>
 #include <thread>
 
 using namespace ELITE;
+namespace po = boost::program_options;
 
 static std::unique_ptr<EliteDriver> s_driver;
 static std::unique_ptr<RtsiIOInterface> s_rtsi_client;
 static std::unique_ptr<DashboardClient> s_dashboard;
 
 int main(int argc, const char** argv) {
-    if (argc < 2) {
-        std::cout << "Must provide robot ip or local ip. Command like: \"./trajectory_example robot_ip\" or \"./trajectory_example "
-                     "robot_ip local_ip\""
-                  << std::endl;
+    EliteDriverConfig config;
+    // Parser param
+    po::options_description desc(
+        "Usage:\n"
+        "\t./trajectory_example <--robot-ip=ip> [--local-ip=\"\"] [--use-headless-mode=true]\n"
+        "Parameters:");
+    desc.add_options()
+        ("help,h", "Print help message")
+        ("robot-ip", po::value<std::string>(&config.robot_ip)->required(),
+            "\tRequired. IP address of the robot.")
+        ("use-headless-mode", po::value<bool>(&config.headless_mode)->required()->implicit_value(true),
+            "\tRequired. Use headless mode.")
+        ("local-ip", po::value<std::string>(&config.local_ip)->default_value(""),
+            "\tOptional. IP address of the local network interface.");
+
+    po::variables_map vm;
+    try {
+        po::store(po::parse_command_line(argc, argv, desc), vm);
+
+        if (vm.count("help")) {
+            std::cout << desc << std::endl;
+            return 0;
+        }
+
+        po::notify(vm);
+    } catch (const po::error& e) {
+        std::cerr << "Argument error: " << e.what() << "\n\n";
+        std::cerr << desc << "\n";
         return 1;
     }
-    std::string robot_ip = argv[1];
-    std::string local_ip = "";
-    if (argc >= 3) {
-        local_ip = argv[2];
+
+    if (config.headless_mode) {
+        ELITE_LOG_WARN("Use headless mode. Please ensure the robot is not in local mode.");
+    } else {
+        ELITE_LOG_WARN(
+            "It needs to be correctly configured, and the External Control plugin should be inserted into the task tree.");
     }
-    EliteDriverConfig config;
-    config.robot_ip = robot_ip;
+
     config.script_file_path = "external_control.script";
-    config.local_ip = local_ip;
     s_driver = std::make_unique<EliteDriver>(config);
 
     s_rtsi_client = std::make_unique<RtsiIOInterface>("output_recipe.txt", "input_recipe.txt", 250);
     s_dashboard = std::make_unique<DashboardClient>();
 
-    if (!s_dashboard->connect(argv[1])) {
+    ELITE_LOG_INFO("Connecting to the dashboard");
+    if (!s_dashboard->connect(config.robot_ip)) {
+        ELITE_LOG_FATAL("Failed to connect to the dashboard.");
         return 1;
     }
-    std::cout << "Dashboard connected" << std::endl;
+    ELITE_LOG_INFO("Successfully connected to the dashboard");
 
-    s_rtsi_client->connect(argv[1]);
-    std::cout << "RTSI connected" << std::endl;
+    ELITE_LOG_INFO("Connecting to the RTSI");
+    if (!s_rtsi_client->connect(config.robot_ip)) {
+        ELITE_LOG_FATAL("Fail to connect or config to the RTSI");
+        return 1;
+    }
+    ELITE_LOG_INFO("Successfully connected to the RTSI");
 
     bool is_move_finish = false;
     s_driver->setTrajectoryResultCallback([&](TrajectoryMotionResult result) {
@@ -49,24 +82,38 @@ int main(int argc, const char** argv) {
         }
     });
 
+    ELITE_LOG_INFO("Start powering on...");
     if (!s_dashboard->powerOn()) {
+        ELITE_LOG_FATAL("Power-on failed");
         return 1;
     }
-    std::cout << "Robot power on" << std::endl;
+    ELITE_LOG_INFO("Power-on succeeded");
 
+    ELITE_LOG_INFO("Start releasing brake...");
     if (!s_dashboard->brakeRelease()) {
         return 1;
     }
-    std::cout << "Robot brake released" << std::endl;
+    ELITE_LOG_INFO("Brake released");
 
-    if (!s_dashboard->playProgram()) {
-        return 1;
+    if (config.headless_mode) {
+        if (!s_driver->isRobotConnected()) {
+            if (!s_driver->sendExternalControlScript()) {
+                ELITE_LOG_FATAL("Fail to send external control script");
+                return 1;
+            }
+        }
+    } else {
+        if (!config.headless_mode && !s_dashboard->playProgram()) {
+            ELITE_LOG_FATAL("Fail to play program");
+            return 1;
+        }
     }
-    std::cout << "Program run" << std::endl;
 
+    ELITE_LOG_INFO("Wait external control script run...");
     while (!s_driver->isRobotConnected()) {
         std::this_thread::sleep_for(std::chrono::milliseconds(10));
     }
+    ELITE_LOG_INFO("External control script is running");
 
     vector6d_t actual_joints = s_rtsi_client->getActualJointPositions();
 
@@ -79,7 +126,7 @@ int main(int argc, const char** argv) {
         s_driver->writeTrajectoryControlAction(ELITE::TrajectoryControlAction::NOOP, 0, 200);
         std::this_thread::sleep_for(std::chrono::milliseconds(100));
     }
-    std::cout << "Joints moved to target" << std::endl;
+   ELITE_LOG_INFO("Joints moved to target");
 
     vector6d_t actual_pose = s_rtsi_client->getAcutalTCPPose();
 
@@ -101,7 +148,7 @@ int main(int argc, const char** argv) {
         s_driver->writeTrajectoryControlAction(ELITE::TrajectoryControlAction::NOOP, 0, 200);
         std::this_thread::sleep_for(std::chrono::milliseconds(100));
     }
-    std::cout << "Line moved to target" << std::endl;
+    ELITE_LOG_INFO("Line moved to target");
 
     s_driver->writeTrajectoryControlAction(ELITE::TrajectoryControlAction::CANCEL, 1, 200);
     s_driver->stopControl();